+++ .gitignore
@@ -2,4 +2,5 @@
 /node_modules/
 yarn-error.log
 **/*.js
+!/misc/**/*.js
 !/src/**/*.js
