+++ package.json
@@ -23,6 +23,7 @@
   "devDependencies": {
     "babel-cli": "^6.24.1",
     "babel-eslint": "^7.2.3",
+    "babel-plugin-module-resolver": "^2.7.1",
     "babel-preset-env": "^1.6.0",
     "babel-preset-stage-2": "^6.24.1",
     "codecov": "^2.3.0",
@@ -36,12 +37,15 @@
   },
   "scripts": {
     "build": "BABEL_ENV=production babel src -d .",
-    "lint": "eslint src",
+    "lint": "eslint src misc",
     "test": "jest",
     "test:coverage": "jest --coverage && codecov",
     "docs": "jsdoc -c jsdoc.json",
     "docs:private": "jsdoc -c jsdoc.json -p"
   },
   "runkitExample": "require('lodash')\nconst immutadot = require('immutadot')\n\nimmutadot.push({ nested: { prop: [1, 2] } }, 'nested.prop', 3, 4)\n// → { nested: { prop: [1, 2, 3, 4] } }\n\npickBy({ nested: [{ a: 1, b: 2, c: 3, d: 4 }, { e: 6 }] }, 'nested.0', v => v < 3)\n// → { nested: [{ a: 1, b: 2 }, { e: 6 }] }",
-  "dependencies": {}
+  "dependencies": {},
+  "jest": {
+    "setupTestFrameworkScriptFile": "./misc/test.setup.js"
+  }
 }
