+++ src/array/takeRight.js
@@ -1,5 +1,5 @@
 import _takeRight from 'lodash/takeRight'
-import { convert } from '../util/convert'
+import { convert } from 'util/convert'
 
 /**
  * Creates a slice of array with <code>n</code> elements taken from the end.
