+++ src/collection/map.js
@@ -1,5 +1,5 @@
 import _map from 'lodash/map'
-import { convert } from '../util/convert'
+import { convert } from 'util/convert'
 
 /**
  * Replaces by an array of values by running each element in the former collection thru iteratee.
