+++ src/math/add.js
@@ -1,5 +1,5 @@
 import _add from 'lodash/add'
-import { convert } from '../util/convert'
+import { convert } from 'util/convert'
 
 /**
  * Replaces by the addition of the former number and the given number.
