+++ src/object/assign.js
@@ -1,5 +1,5 @@
 import _assign from 'lodash/assign'
-import { convert } from '../util/convert'
+import { convert } from 'util/convert'
 
 /**
  * Assigns own enumerable string keyed properties of source objects to the
