+++ src/object/mapValues.js
@@ -1,5 +1,5 @@
 import _mapValues from 'lodash/mapValues'
-import { convert } from '../util/convert'
+import { convert } from 'util/convert'
 
 /**
  * Creates an object with the same keys as object and values generated by
